Refactor *nix code into DFInstanceNix, misc changes
